Overhauls authenticator architecture
We previously treated HMAC based JWTs as second class
citizens and only used them for local testing, otherwise
relying completely on JWKS.

We now support both to be present. If an incoming key has
a "kid" in its header and a JWKS url is present, we
authenticate against that. Otherwise, we try to
verify based on the HMAC signature of a given shared
secret (if present).

Additionally, we permit to supply additional config
fetchers. These receive the (unverified) JWT claims and
can then select an authentication configuration (e.g.
based on a tenant id). If none of the fetchers yields
a configuration, the system config (as given in the
environment) is used.